apps/api-fastapi/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.27.2
SQLAlchemy==2.0.36
aiosqlite==0.20.0
greenlet==3.1.1
celery==5.3.6
redis==5.0.1
Pillow==10.4.0
